Added an icon for future Analysis mode. Added a highlight to the variables list. Then did a ton of cleaning/organizing of the UI folder
