+++ src/components/LandingHome.tsx
@@ -6,9 +6,9 @@ import { ZarrDataset, GetStore } from '@/components/zarr/ZarrLoaderLRU';
 import { useRef, useState } from 'react';
 import VariableScroller from './ui/VariableScroller';
 import { useEffect, useMemo } from 'react';
-import { Analysis, PlotArea, Plot } from '@/components/plots';
-import { MiddleSlider, MainPanel } from '@/components/ui';
-import { Metadata, ShowAnalysis, Loading, Navbar, ShowPlot, Error } from '@/components/ui';
+import { PlotArea, Plot } from '@/components/plots';
+import { MainPanel } from '@/components/ui';
+import { Metadata, Loading, Navbar, Error } from '@/components/ui';
 import { useGlobalStore, useZarrStore } from '@/utils/GlobalStates';
 import { useShallow, shallow } from 'zustand/shallow';
 import useCSSVariable from '@/components/ui/useCSSVariable';
@@ -62,29 +62,9 @@ export function LandingHome() {
     return () => window.removeEventListener('resize', handleResize);
   }, []);
 
-  //Timeseries Plotting Information
-  const [canvasWidth, setCanvasWidth] = useState<number>(0)
-  useEffect(() => {
-    setCanvasWidth(Math.round(window.innerWidth * 0.0))
-  }, [])
-
-  //These values are passed to the Plot Component
-  const plotObj = useMemo(() => ({
-    ZarrDS,
-    canvasWidth
-  }), [ ZarrDS, canvasWidth]);
-
-//This is the data being passed down the plot tree
-  const analysisObj = useMemo(() => ({
-    values: {
-      ZarrDS,
-      canvasWidth,
-    }
-  }), [ZarrDS, canvasWidth]);
-
   useEffect(()=>{
     if (variable === "Default"){
-      setCanvasWidth(0);
+
       setPlotOn(false)
     }
   },[variable])
@@ -94,15 +74,9 @@ export function LandingHome() {
     <MainPanel/> 
     <Error />
     {!plotOn && <Navbar />}
-    {canvasWidth < 15 && variable != "Default" && <ShowAnalysis onClick={()=>setCanvasWidth(windowWidth*.5)} />}
-    {canvasWidth > windowWidth-15 && variable != "Default" && 
-    <ShowPlot onClick={()=>setCanvasWidth(windowWidth*.5)} />}
-    {canvasWidth > 15 && canvasWidth < windowWidth-15 && 
-      <MiddleSlider canvasWidth={canvasWidth} setCanvasWidth={setCanvasWidth}/>}
     <Loading showLoading={showLoading} />
-    {canvasWidth > 10 && variable != "Default" && <Analysis values={analysisObj.values} />}
     {variable === "Default" && <VariableScroller />}
-    {variable != "Default" && canvasWidth < windowWidth-15 && <Plot values={plotObj} setShowLoading={setShowLoading} />}
+    {variable != "Default" && <Plot ZarrDS={ZarrDS} setShowLoading={setShowLoading} />}
     {metadata && <Metadata data={metadata} /> }
     {Object.keys(timeSeries).length >= 1 && <PlotArea />}
     </>
