+++ src/components/ui/MainPanel/Variables.tsx
@@ -4,7 +4,6 @@ import React, { useState, useEffect } from "react";
 import { TbVariable } from "react-icons/tb";
 import { useGlobalStore } from "@/utils/GlobalStates";
 import { useShallow } from "zustand/shallow";
-import { ScrollArea } from "@/components/ui/scroll-area";
 import { Separator } from "@/components/ui/separator";
 import MetaDataInfo from "./MetaDataInfo";
 import { Popover, PopoverTrigger, PopoverContent } from "@/components/ui/popover"
@@ -13,19 +12,18 @@ const Variables = () => {
   const [popoverSide, setPopoverSide] = useState<"left" | "top">("left");
 
   const [showMeta, setShowMeta] = useState(false);
-  const { variables, zMeta, setVariable } = useGlobalStore(
+  const { variables, zMeta } = useGlobalStore(
     useShallow((state) => ({
       variables: state.variables,
       zMeta: state.zMeta,
-      setVariable: state.setVariable,
     }))
   );
 
-  const [selectedIndex, setSelectedIndex] = useState(0);
+  const [selectedIndex, setSelectedIndex] = useState<number | null>(null);
   const [meta, setMeta] = useState<any>(null);
 
   useEffect(() => {
-    if (variables && zMeta) {
+    if (variables && zMeta && selectedIndex) {
       const tempVar = variables[selectedIndex];
       const relevant = zMeta.find((e: any) => e.name === tempVar);
       setMeta(relevant);
@@ -53,7 +51,8 @@ const Variables = () => {
             {variables.map((val, idx) => (
               <React.Fragment key={idx}>
                 <div
-                  className="cursor-pointer px-2 py-1 text-sm hover:bg-muted rounded"
+                  className="cursor-pointer pl-2 py-1 text-sm hover:bg-muted rounded"
+                  style={{background: idx == selectedIndex ? '#d6d6d6ff' : ''}}
                   onClick={() => {
                     setSelectedIndex(idx);
                     setShowMeta(true);
